first set of big fixes

+++ src/commands/organization/registry.ts
@@ -1,8 +1,9 @@
-import { CommandInteraction } from "discord.js";
+import { CommandInteraction, Permissions } from "discord.js";
 import { SlashCommandBuilder } from '@discordjs/builders';
 import createRegistry from "./subcommands/createRegistry";
 import clearRegistry from "./subcommands/clearRegistry";
 import destroyRegistry from "./subcommands/destroyRegistry";
+import hasPermission from "./helpers/hasPermission";
 
 export default {
 	data: new SlashCommandBuilder()
@@ -53,18 +54,26 @@ export default {
             ),
 
 	async execute(interaction: CommandInteraction) {
-        switch (interaction.options.getSubcommand()) {
-            case 'create':
-                await createRegistry(interaction)
-                break;
-        
-            case 'clear':
-                await clearRegistry(interaction)
-                break;
-
-            case 'destroy':
-                await destroyRegistry(interaction)
-                break;
+        if (interaction.channel?.type === 'GUILD_TEXT') {    
+            if (hasPermission(interaction.member!.permissions, Permissions.FLAGS.ADMINISTRATOR)) {
+                switch (interaction.options.getSubcommand()) {
+                    case 'create':
+                        await createRegistry(interaction)
+                        break;
+                
+                    case 'clear':
+                        await clearRegistry(interaction)
+                        break;
+    
+                    case 'destroy':
+                        await destroyRegistry(interaction)
+                        break;
+                }
+            } else {
+                await interaction.reply({content: 'Insufficient permissions', ephemeral: true})
+            }
+        } else {
+            await interaction.reply({content: 'This command isn\'t functional in dms', ephemeral: true})
         }
 	},
 };
